+++ packages/compass-user-data/.depcheckrc
@@ -1,0 +1,6 @@
+ignores:
+ - "@mongodb-js/prettier-config-compass"
+ - "@mongodb-js/tsconfig-compass"
+ - "@types/chai"
+ - "@types/sinon-chai"
+ - "sinon"
