+++ packages/compass-user-data/.prettierignore
@@ -1,0 +1,2 @@
+.nyc-output
+dist
