+++ packages/compass-user-data/package.json
@@ -1,0 +1,71 @@
+{
+  "name": "@mongodb-js/compass-user-data",
+  "description": "UserData storage helper for MongoDB Compass",
+  "author": {
+    "name": "MongoDB Inc",
+    "email": "compass@mongodb.com"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "bugs": {
+    "url": "https://jira.mongodb.org/projects/COMPASS/issues",
+    "email": "compass@mongodb.com"
+  },
+  "homepage": "https://github.com/mongodb-js/compass",
+  "version": "0.1.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mongodb-js/compass.git"
+  },
+  "files": [
+    "dist"
+  ],
+  "license": "SSPL",
+  "main": "dist/index.js",
+  "exports": {
+    "require": "./dist/index.js",
+    "import": "./dist/.esm-wrapper.mjs"
+  },
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "bootstrap": "npm run compile",
+    "prepublishOnly": "npm run compile",
+    "compile": "tsc -p tsconfig.json && gen-esm-wrapper . ./dist/.esm-wrapper.mjs",
+    "eslint": "eslint",
+    "prettier": "prettier",
+    "lint": "npm run eslint . && npm run prettier -- --check .",
+    "depcheck": "depcheck",
+    "check": "npm run lint && npm run depcheck",
+    "check-ci": "npm run check",
+    "test": "mocha",
+    "test-cov": "nyc -x \"**/*.spec.*\" npm run test",
+    "test-watch": "npm run test -- --watch",
+    "test-ci": "npm run test-cov",
+    "reformat": "npm run prettier -- --write ."
+  },
+  "dependencies": {
+    "@mongodb-js/compass-logging": "^0.3.0",
+    "glob": "^7.2.0",
+    "write-file-atomic": "^3.0.3"
+  },
+  "devDependencies": {
+    "@mongodb-js/eslint-config-compass": "^0.3.0",
+    "@mongodb-js/mocha-config-compass": "^0.4.0",
+    "@mongodb-js/prettier-config-compass": "^0.2.0",
+    "@mongodb-js/tsconfig-compass": "^0.2.0",
+    "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
+    "@types/sinon-chai": "^3.2.5",
+    "@types/write-file-atomic": "^3.0.2",
+    "chai": "^4.3.4",
+    "electron": "^6.1.12",
+    "depcheck": "^1.4.1",
+    "eslint": "^7.25.0",
+    "mocha": "^8.4.0",
+    "nyc": "^15.1.0",
+    "prettier": "2.3.2",
+    "sinon": "^9.2.3",
+    "typescript": "^4.3.5"
+  }
+}
