+++ packages/compass-user-data/src/index.ts
@@ -1,0 +1,1 @@
+export { UserData } from './user-data';
