+++ packages/data-service/package.json
@@ -61,7 +61,6 @@
     "@mongodb-js/ssh-tunnel": "^1.2.0",
     "async": "^3.2.0",
     "debug": "^4.2.0",
-    "glob": "^7.2.0",
     "lodash": "^4.17.20",
     "mongodb-build-info": "^1.1.1",
     "mongodb-connection-string-url": "^2.1.0",
@@ -70,8 +69,7 @@
     "mongodb-ns": "^2.2.0",
     "mongodb-security": "^1.3.0",
     "resolve-mongodb-srv": "^1.1.0",
-    "uuid": "^8.3.2",
-    "write-file-atomic": "^3.0.3"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@mongodb-js/devtools-docker-test-envs": "^1.0.1",
@@ -84,7 +82,6 @@
     "@types/lodash": "^4.14.172",
     "@types/uuid": "^8.3.1",
     "@types/whatwg-url": "^8.2.1",
-    "@types/write-file-atomic": "^3.0.2",
     "bson": "^4.4.1",
     "chai": "^4.2.0",
     "depcheck": "^1.4.1",
